requirements.txt
flask                  3.0.0               
flask-marshmallow      0.15.0   
pymongo                4.5.0  
python-dotenv          1.0.0               
